refactor: ゲーム部分をコンポーネントに分離、ファイルを整理
